pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kiroforge"
version = "0.1.0"
description = "Authoring, validation, and testing toolkit for Kiro Powers."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "KiroForge" }
]
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "pytest>=9.0.2",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2"
]

[project.scripts]
kiroforge = "kiroforge.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
